Fix templates

+++ src/templates/basic/servers/typescript/task.ts
@@ -1,11 +1,11 @@
 import mongoose from "mongoose"
-import { MONGO_DB_URI } from "./helper"
+import { reqAuth, MONGO_DB_URI } from "./helper"
 import { TaskModel } from "./models/task"
 
 export type Task = {
   _id: string,
   title: string,
-  token: string,
+  ownerId: string,
   solved: boolean,
   date: Date,
 }
@@ -18,14 +18,6 @@ export type GetTasksResponse = {
 export type GetTaskResponse = {
   success: boolean,
   task?: Task
-}
-
-export type CreateTaskResponse = {
-  success: boolean,
-  task: {
-    title: string,
-    _id: string,
-  }
 }
 
 export type UpdateTaskResponse = {
@@ -48,88 +40,72 @@ export class TaskService {
    * Private method used to connect to the DB.
    */
   #connect() {
-    mongoose.set('strictQuery', true);
     mongoose.connect(MONGO_DB_URI);
   }
 
   /**
    * Method that returns all tasks for a giving user ID.
    * Only authenticated users with a valid token can access this method.
-   *
+   * 
    * The method will be exported via SDK using genezio.
-   *
+   * 
    * @param {*} token The user's token.
    * @param {*} userId The user ID.
    * @returns An object containing two properties: { success: true, tasks: tasks }
    */
-  async getAllTasksByUser(token: string): Promise<GetTasksResponse> {
-    console.log(`Get all tasks by user request received with token ${token}`)
+  async getAllTasksByUser(token: string, userId: string): Promise<GetTasksResponse> {
+    console.log(`Get all tasks by user request received with userID ${userId}`)
 
-    const tasks = await TaskModel.find({ token: token });
-
-    if (tasks.length === 0) {
-      await TaskModel.create({
-        token: token,
-        title: "Check the other example projects",
-        url: "https://github.com/Genez-io/genezio-examples"
-      })
-
-      await TaskModel.create({
-        token: token,
-        title: "Check our documentation",
-        url: "https://docs.genez.io/genezio-documentation/"
-      })
-
-      await TaskModel.create({
-        token: token,
-        title: "Watch our Youtube tutorials",
-        url: "https://www.youtube.com/@genezio7235"
-      })
-
-      await TaskModel.create({
-        token: token,
-        title: "Read our technical articles on genezio blog",
-        url: "https://genez.io/blog/"
-      })
-
-      const initTasks = await TaskModel.find({ token: token });
-
-      return { success: true, tasks: initTasks };
+    const authObject = await reqAuth(token);
+    if (!authObject.success) {
+      return { success: false, tasks: [] };
     }
-
+    const tasks = (await TaskModel.find({ ownerId: userId }))
+      .map((task) => ({
+        title: task.title,
+        ownerId: task.ownerId,
+        solved: task.solved,
+        date: task.date,
+        _id: task._id.toString()
+      }));
     return { success: true, tasks: tasks };
   }
 
   /**
    * Method that creates a task for a giving user ID.
    * Only authenticated users with a valid token can access this method.
-   *
+   * 
    * The method will be exported via SDK using genezio.
-   *
+   * 
    * @param {*} token The user's token.
    * @param {*} title The tasktitle.
+   * @param {*} ownerId The owner's of the task ID.
    * @returns An object containing two properties: { success: true, tasks: tasks }
    */
-  async createTask(token:string, title:string) : Promise<CreateTaskResponse> {
-    console.log(`Create task request received for user with ${token} with title ${title}`)
+  async createTask(token: string, title: string, ownerId: string): Promise<GetTaskResponse> {
+    console.log(`Create task request received for user with id ${ownerId} with title ${title}`)
 
+    const authObject = await reqAuth(token);
+    if (!authObject.success) {
+      return { success: false };
+    }
     const task = await TaskModel.create({
       title: title,
-      token: token
+      ownerId: ownerId
     });
 
     return {
       success: true,
-      task: { title: title, _id: task._id.toString() }
+      task: { title: title, ownerId: ownerId, _id: task._id.toString(), solved: false, date: new Date() }
     };
   }
 
   /**
    * Method that creates a task for a giving user ID.
    * Only authenticated users with a valid token can access this method.
-   *
+   * 
    * The method will be exported via SDK using genezio.
-   *
+   * 
    * @param {*} token The user's token.
    * @param {*} id The task's id.
    * @param {*} title The task's title.
@@ -139,6 +115,10 @@ export class TaskService {
   async updateTask(token: string, id: string, title: string, solved: boolean): Promise<UpdateTaskResponse> {
     console.log(`Update task request received with id ${id} with title ${title} and solved value ${solved}`)
 
+    const authObject = await reqAuth(token);
+    if (!authObject.success) {
+      return authObject;
+    }
     await TaskModel.updateOne(
       { _id: id },
       {
@@ -153,9 +133,9 @@ export class TaskService {
   /**
    * Method that deletes a task for a giving user ID.
    * Only authenticated users with a valid token can access this method.
-   *
+   * 
    * The method will be exported via SDK using genezio.
-   *
+   * 
    * @param {*} token The user's token.
    * @param {*} title The tasktitle.
    * @returns An object containing one property: { success: true }
@@ -163,6 +143,10 @@ export class TaskService {
   async deleteTask(token: string, id: string): Promise<DeleteTaskResponse> {
     console.log(`Delete task with id ${id} request received`)
 
+    const authObject = await reqAuth(token);
+    if (!authObject.success) {
+      return authObject;
+    }
     await TaskModel.deleteOne({ _id: id });
 
     return { success: true };
